SSP-2: refactoring of cypress tests
