+++ cypress/fixtures/suite/backoffice/ssp-file-management/dynamic-ssp-file-management.json
