+++ cypress/fixtures/suite/backoffice/ssp-file-management/static-ssp-file-management.json
