+++ cypress/fixtures/suite/backoffice/ssp-file-management/test-files/document1.pdf
