+++ cypress/fixtures/suite/backoffice/ssp-file-management/test-files/document2.pdf
