+++ cypress/fixtures/suite/backoffice/ssp-file-management/test-files/image1.jpeg
