+++ cypress/fixtures/suite/backoffice/ssp-file-management/test-files/image2.png
