--- cypress/fixtures/suite/yves/file-manager-attachment/static-file-manager-attachment.json
@@ -1,3 +1,0 @@
-{
-    "defaultPassword": "change123"
-}
