+++ cypress/fixtures/suite/yves/ssp-file-management/dynamic-ssp-file-management.json
@@ -1,0 +1,201 @@
+{
+  "data": {
+    "type": "dynamic-fixtures",
+    "attributes": {
+      "synchronize": true,
+      "operations": [
+        {
+          "type": "transfer",
+          "name": "LocaleTransfer",
+          "key": "localeEN",
+          "arguments": { "id_locale": 66, "locale_name": "en_US" }
+        },
+        {
+          "type": "transfer",
+          "name": "LocaleTransfer",
+          "key": "localeDE",
+          "arguments": { "id_locale": 46, "locale_name": "de_DE" }
+        },
+        {
+          "type": "builder",
+          "name": "LocalizedAttributesBuilder",
+          "key": "localizedAttribute"
+        },
+        {
+          "type": "transfer",
+          "name": "StoreTransfer",
+          "key": "store",
+          "arguments": { "id_store": 1, "name": "DE" }
+        },
+        {
+          "type": "helper",
+          "name": "haveCompany",
+          "key": "company1",
+          "arguments": [{ "isActive": true, "status": "approved", "name": "Test Company 1" }]
+        },
+        {
+          "type": "helper",
+          "name": "haveCompany",
+          "key": "company2",
+          "arguments": [{ "isActive": true, "status": "approved", "name": "Test Company 2" }]
+        },
+        {
+          "type": "helper",
+          "name": "haveCompany",
+          "key": "company3",
+          "arguments": [{ "isActive": true, "status": "approved", "name": "Test Company 3" }]
+        },
+        {
+          "type": "helper",
+          "name": "haveCompanyBusinessUnit",
+          "key": "businessUnit",
+          "arguments": [
+            {
+              "fkCompany": "#company1.id_company",
+              "name": "Test Business Unit"
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveCustomer",
+          "key": "customer",
+          "arguments": [
+            {
+              "locale": "#localeEN",
+              "password": "change123",
+              "first_name": "Test",
+              "last_name": "Customer"
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "confirmCustomer",
+          "key": "confirmedCustomer",
+          "arguments": ["#customer"]
+        },
+        {
+          "type": "helper",
+          "name": "haveCompanyUser",
+          "key": "companyUser",
+          "arguments": [
+            {
+              "customer": "#customer",
+              "fkCustomer": "#customer.id_customer",
+              "fkCompany": "#company1.id_company",
+              "fkCompanyBusinessUnit": "#businessUnit.id_company_business_unit"
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "havePermissionByKey",
+          "key": "permission",
+          "arguments": ["ViewFilesPermissionPlugin"]
+        },
+        {
+          "type": "helper",
+          "name": "havePermissionByKey",
+          "key": "permission2",
+          "arguments": ["DownloadFilesPermissionPlugin"]
+        },
+        {
+          "type": "helper",
+          "name": "havePermissionByKey",
+          "key": "permission3",
+          "arguments": ["ViewCompanyUserFilesPermissionPlugin"]
+        },
+        {
+          "type": "helper",
+          "name": "haveCompanyRoleWithPermissions",
+          "key": "companyRoleWithAllPermissions",
+          "arguments": [
+            { "isDefault": true, "fkCompany": "#company1.id_company" },
+            ["#permission", "#permission2", "#permission3"]
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "assignCompanyRoles",
+          "arguments": ["#companyUser", ["#companyRoleWithAllPermissions"]]
+        },
+        {
+          "type": "helper",
+          "name": "haveFile",
+          "key": "file1",
+          "arguments": [
+            {
+              "file_name": "img.jpeg"
+            },
+            {
+              "extension": "jpeg"
+            },
+            "ssp-files"
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveFile",
+          "key": "file2",
+          "arguments": [
+            {
+              "file_name": "doc.pdf"
+            },
+            {
+              "extension": "pdf"
+            },
+            "ssp-files"
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveFile",
+          "key": "file3",
+          "arguments": [
+            {
+              "file_name": "file.pdf"
+            },
+            {
+              "extension": "pdf"
+            },
+            "ssp-files"
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveFileAttachment",
+          "arguments": [
+            {
+              "file": "#file1",
+              "entity_id": "#companyUser.id_company_user",
+              "entity_name": "company_user"
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveFileAttachment",
+          "arguments": [
+            {
+              "file": "#file2",
+              "entity_id": "#companyUser.id_company_user",
+              "entity_name": "company_user"
+            }
+          ]
+        },
+        {
+          "type": "helper",
+          "name": "haveFileAttachment",
+          "arguments": [
+            {
+              "file": "#file3",
+              "entity_id": "#companyUser.id_company_user",
+              "entity_name": "company_user"
+            }
+          ]
+        }
+      ]
+    }
+  }
+}
