+++ cypress/fixtures/suite/yves/ssp-file-management/static-ssp-file-management.json
@@ -1,0 +1,9 @@
+{
+    "defaultPassword": "change123",
+    "filter_type_file": "file",
+    "filter_value_pdf": "pdf",
+    "filter_value_jpeg": "jpeg",
+    "prompt_img": "img",
+    "prompt_doc": "doc",
+    "prompt_nonexistent": "nonexistent"
+}
