--- cypress/support/pages/backoffice/file-manager-attachment/add/file-manager-attachment-add-repository.ts
@@ -1,10 +1,0 @@
-import { injectable } from 'inversify';
-import { autoWired } from '@utils';
-
-@injectable()
-@autoWired
-export class FileManagerAttachmentAddRepository {
-  getFileInputSelector = (): string => '[data-qa="file-upload-button"]';
-  getSubmitButtonSelector = (): string => '[data-qa="file-upload-submit"]';
-  getSuccessMessageSelector = (): string => '[data-qa="success-message"]';
-}
