--- cypress/support/pages/backoffice/file-manager-attachment/detach/file-manager-attachment-detach-page.ts
@@ -1,24 +1,0 @@
-import { injectable, inject } from 'inversify';
-import { autoWired } from '@utils';
-import { BackofficePage } from '@pages/backoffice';
-import { FileManagerAttachmentDetachRepository } from './file-manager-attachment-detach-repository';
-
-@injectable()
-@autoWired
-export class FileManagerAttachmentDetachPage extends BackofficePage {
-  @inject(FileManagerAttachmentDetachRepository) private repository: FileManagerAttachmentDetachRepository;
-
-  detachFile(): void {
-    cy.get(this.repository.getDetachButtonSelector()).first().click();
-  }
-
-  assertDetachFile(): void {
-    this.repository.getAttachmentTableRows().should('have.length.greaterThan', 1);
-  }
-
-  verifySuccessMessage(): void {
-    cy.get(this.repository.getSuccessMessageSelector())
-      .should('be.visible')
-      .and('contain', 'File attachment successfully unlinked.');
-  }
-}
