+++ cypress/support/pages/backoffice/index.ts
@@ -36,11 +36,11 @@ export * from './shipment/shipment-method/list/shipment-method-list-page';
 export * from './shipment/shipment-method/edit/shipment-method-edit-page';
 export * from './payment-method/list/payment-method-list-page';
 export * from './payment-method/edit/payment-method-edit-page';
-export * from './file-manager-attachment/list/file-manager-attachment-list-page';
-export * from './file-manager-attachment/add/file-manager-attachment-add-page';
-export * from './file-manager-attachment/view/file-manager-attachment-view-page';
-export * from './file-manager-attachment/delete/file-manager-attachment-delete-page';
-export * from './file-manager-attachment/detach/file-manager-attachment-detach-page';
-export * from './file-manager-attachment/attach/file-manager-attachment-attach-page';
+export * from './ssp-file-management/list/ssp-file-management-list-page';
+export * from './ssp-file-management/add/ssp-file-management-add-page';
+export * from './ssp-file-management/view/ssp-file-management-view-page';
+export * from './ssp-file-management/delete/ssp-file-management-delete-page';
+export * from './ssp-file-management/detach/ssp-file-management-detach-page';
+export * from './ssp-file-management/attach/ssp-file-management-attach-page';
 export * from './company-role/list/company-role-list-page';
 export * from './company-role/create/company-role-create-page';
