+++ cypress/support/pages/backoffice/ssp-file-management/attach/ssp-file-management-attach-repository.ts
@@ -1,0 +1,19 @@
+import { injectable } from 'inversify';
+import { autoWired } from '@utils';
+
+@injectable()
+@autoWired
+export class SspFileManagementAttachRepository {
+  getAttachButtonSelector = (): string => '[data-qa="attach-button"]';
+  getCompanyFieldSelector = (): string => '[data-qa="attach-company-field"]';
+  getCompanyUserFieldSelector = (): string => '[data-qa="attach-company-user-field"]';
+  getCompanyBusinessUnitFieldSelector = (): string => '[data-qa="attach-company-business-unit-field"]';
+  getSiblingSelector = (): string => 'span';
+  getSearchFieldSelector = (): string => 'input.select2-search__field';
+  getDropdownOptionSelector = (): string => '.select2-results__option';
+  getSubmitButtonSelector = (): string => '[data-qa="attach-submit-button"]';
+  getSuccessMessageSelector = (): string => '[data-qa="success-message"]';
+  getCompanyPrompt = (): string => 'test';
+  getCompanyUserPrompt = (): string => 'test';
+  getCompanyBusinessUnitPrompt = (): string => 'test';
+}
