+++ cypress/support/pages/backoffice/ssp-file-management/delete/ssp-file-management-delete-repository.ts
@@ -1,0 +1,11 @@
+import { injectable } from 'inversify';
+import { autoWired } from '@utils';
+import Chainable = Cypress.Chainable;
+
+@injectable()
+@autoWired
+export class SspFileManagementDeleteRepository {
+  getDeleteConfirmButtonSelector = (): string => '[data-qa="delete-confirm-button"]';
+  getSuccessMessageSelector = (): string => '[data-qa="success-message"]';
+  getFileTableRows = (): Chainable => cy.get('table.gui-table-data tbody tr');
+}
