+++ cypress/support/pages/backoffice/ssp-file-management/detach/ssp-file-management-detach-repository.ts
@@ -1,0 +1,11 @@
+import { injectable } from 'inversify';
+import { autoWired } from '@utils';
+import Chainable = Cypress.Chainable;
+
+@injectable()
+@autoWired
+export class SspFileManagementDetachRepository {
+  getDetachButtonSelector = (): string => '[data-qa="unlink-button"]';
+  getSuccessMessageSelector = (): string => '[data-qa="success-message"]';
+  getAttachmentTableRows = (): Chainable => cy.get('table.gui-table-data tbody tr');
+}
