+++ cypress/support/pages/yves/customer/overview/repositories/b2c-mp-customer-overview-repository.ts
@@ -14,4 +14,7 @@ export class B2cMpCustomerOverviewRepository implements CustomerOverviewReposito
   getViewOrderButton(tableRowIndex: number): Cypress.Chainable {
     return cy.get('[data-qa="component order-table"]').find('tr').eq(tableRowIndex).contains('a', 'View Order');
   }
+  getMyFilesLink(): Cypress.Chainable {
+    return cy.get('[data-qa="my-files"]');
+  }
 }
