+++ cypress/support/pages/yves/index.ts
@@ -23,4 +23,4 @@ export * from './product-comparison/product-comparison-page';
 export * from './content/cms-content-page';
 export * from './company/company-role/index/company-role-index-page';
 export * from './company/company-role/create/company-role-create-page';
-export * from './file-manager-attachment/list/file-manager-attachment-list-page';
+export * from './ssp-file-management/list/ssp-file-management-list-page';
