+++ cypress/support/pages/yves/ssp-file-management/list/ssp-file-management-list-repository.ts
@@ -1,0 +1,17 @@
+import { injectable } from 'inversify';
+import { autoWired } from '@utils';
+
+@injectable()
+@autoWired
+export class SspFileManagementListRepository {
+    getFitersSelector = (): string => '[data-qa="component filters"]';
+    getFileTableSelector = (): string => '[data-qa="component view-table"]';
+    getFileNameHeaderSelector = (): string => '[data-qa="name"]';
+    getUploadedDateHeaderSelector = (): string => '[data-qa="date"]';
+    getFileSizeHeaderSelector = (): string => '[data-qa="size"]';
+    getFileTypeHeaderSelector = (): string => '[data-qa="type"]';
+    getDownloadButtonSelector = (): string => '[data-qa="download-button"]';
+    getTypeFilterSelector = (): string => '[data-qa="filter-type"]';
+    getApplyFiltersButtonSelector = (): string => '[data-qa="filters-apply"]';
+    getSearchFieldSelector = (): string => '[data-qa="search"]';
+}
