--- cypress/support/types/backoffice/file-list.ts
@@ -1,12 +1,0 @@
-import { User } from './shared';
-import { File, FileAttachment } from '@interfaces/shared';
-
-export interface FileManagerAttachmentStaticFixtures {
-  defaultPassword: string;
-}
-
-export interface FileManagerAttachmentDynamicFixtures {
-  file: File;
-  fileAttachment: FileAttachment;
-  rootUser: User;
-}
