+++ cypress/support/types/backoffice/index.ts
@@ -3,4 +3,4 @@ export * from './order-management';
 export * from './return-management';
 export * from './merchant-b2b-contract-requests';
 export * from './company-role-create';
-export * from './file-list';
+export * from './ssp-file-management';
