--- cypress/support/types/yves/file-manager-attachment.ts
@@ -1,9 +1,0 @@
-import { Customer } from './shared';
-
-export interface FileManagerAttachmentDynamicFixtures {
-    customer: Customer;
-}
-
-export interface FileManagerAttachmentStaticFixtures {
-    defaultPassword: string;
-}
