+++ cypress/support/types/yves/index.ts
@@ -11,4 +11,4 @@ export * from './order-amendment';
 export * from './locale';
 export * from './catalog';
 export * from './company-role';
-export * from './file-manager-attachment';
+export * from './ssp-file-management';
