+++ cypress/support/types/yves/shared/file-info.ts
@@ -1,0 +1,3 @@
+export interface FileInfo {
+    storage_file_name: string;
+}
