+++ cypress/support/types/yves/shared/file.ts
@@ -1,0 +1,6 @@
+import { FileInfo } from './file-info';
+
+export interface File {
+    file_info: FileInfo[];
+    file_name: string;
+}
